+++ Common/Gurrex.Common/Gurrex.Common.Interfaces/Repositories/IEntityRepository.cs
@@ -1,5 +1,4 @@
 ﻿using Gurrex.Common.Interfaces.Entities;
-using Gurrex.Common.Interfaces.Repositories.Base;
 using System.Linq.Expressions;
 
 namespace Gurrex.Common.Interfaces.Repositories
@@ -8,8 +7,14 @@ namespace Gurrex.Common.Interfaces.Repositories
     /// Асинхронный репозиторий работы с сущностью, реализующей интерфейс <see cref="IEntity"/>
     /// </summary>
     /// <typeparam name="T">Сущность, реализующая интерфейс <see cref="IEntity"/></typeparam>
-    public interface IRepositoryEntitiesAsync<T> : IRepository<T> where T : class, IEntity, new()
+    public interface IEntityRepository<T> where T : class, IEntity, new()
     {
+
+        /// <summary>
+        /// Получить все экземпляры сущности, унаследованной от <see cref="IEntity"/>
+        /// </summary>
+        IQueryable<T> Items { get; }
+
         /// <summary>
         /// Асинхронно получить экземпляр сущности по Id
         /// </summary>
@@ -71,5 +76,45 @@ namespace Gurrex.Common.Interfaces.Repositories
         /// <param name="cancel">Токен отмены</param>
         /// <returns>True - изменения сохранены, False - изменения не сохранены</returns>
         Task<bool> SaveChangesAsync(CancellationToken cancel = default);
+
+        /// <summary>
+        /// Получить экземпляр сущности по Id
+        /// </summary>
+        /// <param name="id">Id сущности</param>
+        /// <returns>Полученный экземпляр сущности</returns>
+        T GetEntityById(Guid id);
+
+        /// <summary>
+        /// Получить последний добавленный экземпляр сущности
+        /// </summary>
+        /// <returns>Последний добавленный экземпляр сущности</returns>
+        T GetLastEntity();
+
+        /// <summary>
+        /// Добавить экземпляр сущности
+        /// </summary>
+        /// <param name="entity">Добавляемый экземпляр сущности</param>
+        /// <returns>True - сущность успешно добавлена, False - сущность не добавлена</returns>
+        bool AddEntity(T entity);
+
+        /// <summary>
+        /// Обновить экземпляр сущности
+        /// </summary>
+        /// <param name="entity">Обновляемый экземпляр сущности</param>
+        /// <returns>True - сущность успешно обновлена, False - сущность не обновлена</returns>
+        bool UpdateEntity(T entity);
+
+        /// <summary>
+        /// Удалить экземпляр сущности по Id
+        /// </summary>
+        /// <param name="id">Id сущности</param>
+        /// <returns>True - сущность удалена, False - сущность не удалена</returns>
+        bool RemoveEntityById(Guid id);
+
+        /// <summary>
+        /// Сохранить изменения в базу
+        /// </summary>
+        /// <returns>True - изменения сохранены, Fales - изменения не сохранены</returns>
+        bool SaveChanges();
     }
 }
