--- Microservices/YouTubeVideoDownloader/YouTubeVideoDownloader.DAL/Repositories/Async/YouTubeInfoRepositoryAsync.cs
@@ -1,25 +1,0 @@
-﻿using Gurrex.Common.DAL.Repositories;
-using Microsoft.Extensions.Logging;
-using YouTubeVideoDownloader.DAL.Context;
-using YouTubeVideoDownloader.DAL.Entities;
-using YouTubeVideoDownloader.Interfaces.DAL.Repositories.Async;
-
-namespace YouTubeVideoDownloader.DAL.Repositories.Async
-{
-    /// <summary>
-    /// Асинхронный репозиторий работы с <see cref="YouTubeInfo"/>
-    /// </summary>
-    public class YouTubeInfoRepositoryAsync : DbRepositoryAsync<YouTubeInfo>, IYouTubeInfoRepositoryAsync<YouTubeInfo>
-    {
-        /// <summary>
-        /// Конструктор инициализатор
-        /// </summary>
-        /// <param name="downloaderContext">Контекст базы данных</param>
-        /// <param name="logger">Логирование</param>
-        public YouTubeInfoRepositoryAsync(DownloaderContext downloaderContext, ILogger<YouTubeInfoRepositoryAsync> logger)
-            : base(downloaderContext, logger)
-        {
-
-        }
-    }
-}
