--- Microservices/YouTubeVideoDownloader/YouTubeVideoDownloader.DAL/Repositories/Sync/AudioRepository.cs
@@ -1,25 +1,0 @@
-﻿using Gurrex.Common.DAL.Repositories;
-using Microsoft.Extensions.Logging;
-using YouTubeVideoDownloader.DAL.Context;
-using YouTubeVideoDownloader.DAL.Entities;
-using YouTubeVideoDownloader.Interfaces.DAL.Repositories.Sync;
-
-namespace YouTubeVideoDownloader.DAL.Repositories.Sync
-{
-    /// <summary>
-    /// Репозиторий работы с <see cref="Audio"/>
-    /// </summary>
-    public class AudioRepository : DbRepository<Audio>, IAudioRepository<Audio>
-    {
-        /// <summary>
-        /// Конструктор инициализатор
-        /// </summary>
-        /// <param name="downloaderContext">Контекст базы данных</param>
-        /// <param name="logger">Логирование</param>
-        public AudioRepository(DownloaderContext downloaderContext, ILogger<AudioRepository> logger) 
-            : base(downloaderContext, logger)
-        {
-
-        }
-    }
-}
