+++ Microservices/YouTubeVideoDownloader/YouTubeVideoDownloader.Interfaces/DAL/IDownloaderRepository.cs
@@ -1,0 +1,15 @@
+﻿using Gurrex.Common.Interfaces.Entities;
+using Gurrex.Common.Interfaces.Repositories;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace YouTubeVideoDownloader.Interfaces.DAL
+{
+    public interface IDownloaderRepository<T> : IEntityRepository<T> where T: class, IEntity, new()
+    {
+
+    }
+}
