+++ Microservices/YouTubeVideoDownloader/YouTubeVideoDownloader.YouTubeDataOperations/Services/ServicesRegistrator.cs
@@ -4,9 +4,7 @@ using Gurrex.Web.SignalR.Hubs.Async;
 using Microsoft.Extensions.DependencyInjection;
 using YouTubeVideoDownloader.DAL;
 using YouTubeVideoDownloader.DAL.Entities;
-using YouTubeVideoDownloader.DAL.Repositories.Async;
 using YouTubeVideoDownloader.Interfaces.DAL;
-using YouTubeVideoDownloader.Interfaces.DAL.Repositories.Async;
 using YouTubeVideoDownloader.Interfaces.Services.Async;
 using YouTubeVideoDownloader.Interfaces.Services.Sync;
 using YouTubeVideoDownloader.YouTubeDataOperations.Models;
